tests were in wrong file
